+++ kinto/plugins/accounts/authentication.py
@@ -5,31 +5,31 @@ from kinto.core import utils
 from kinto.core.storage import exceptions as storage_exceptions
 
 from .utils import (
-    ACCOUNT_CACHE_KEY,
-    ACCOUNT_POLICY_NAME,
-    ACCOUNT_RESET_PASSWORD_CACHE_KEY,
-    is_validated,
+    cache_account,
+    get_account_cache_key,
+    get_cached_account,
     get_cached_reset_password,
+    delete_cached_reset_password,
+    is_validated,
+    refresh_cached_account,
+    ACCOUNT_POLICY_NAME,
 )
 
 
 def account_check(username, password, request):
     settings = request.registry.settings
-    hmac_secret = settings["userid_hmac_secret"]
     validation_enabled = settings.get("account_validation", False)
-    cache_key = utils.hmac_digest(hmac_secret, ACCOUNT_CACHE_KEY.format(username))
-    cache_ttl = int(settings.get("account_cache_ttl_seconds", 30))
+    cache_key = get_account_cache_key(username, request.registry)
     hashed_password = utils.hmac_digest(cache_key, password)
 
     # Check cache to see whether somebody has recently logged in with the same
     # username and password.
-    cache = request.registry.cache
-    cache_result = cache.get(cache_key)
+    cache_result = get_cached_account(username, request.registry)
 
     # Username and password have been verified previously. No need to compare hashes
     if cache_result == hashed_password:
         # Refresh the cache TTL.
-        cache.expire(cache_key, cache_ttl)
+        refresh_cached_account(username, request.registry)
         return True
 
     # Back to standard procedure
@@ -48,21 +48,17 @@ def account_check(username, password, request):
     pwd_str = password.encode(encoding="utf-8")
     # Check if password is valid (it is a very expensive computation)
     if bcrypt.checkpw(pwd_str, hashed):
-        cache.set(cache_key, hashed_password, ttl=cache_ttl)
+        cache_account(hashed_password, username, request.registry)
         return True
 
     # Last chance, is this a "reset password" flow?
-    reset_password_flow(username, password, request)
+    return reset_password_flow(username, password, request)
 
 
 def reset_password_flow(username, password, request):
-    settings = request.registry.settings
-    hmac_secret = settings["userid_hmac_secret"]
-    cache_key = utils.hmac_digest(hmac_secret, ACCOUNT_CACHE_KEY.format(username))
+    cache_key = get_account_cache_key(username, request.registry)
     hashed_password = utils.hmac_digest(cache_key, password)
     pwd_str = password.encode(encoding="utf-8")
-    cache_ttl = int(settings.get("account_cache_ttl_seconds", 30))
-    cache = request.registry.cache
 
     cached_password = get_cached_reset_password(username, request.registry)
     if not cached_password:
@@ -80,7 +76,7 @@ def reset_password_flow(username, password, request):
 
     try:
         data = request.json["data"]
-    except ValueError:
+    except (ValueError, KeyError):
         return None
 
     # Request one and only one data field: the `password`.
@@ -90,11 +86,8 @@ def reset_password_flow(username, password, request):
     cached_password_str = cached_password.encode(encoding="utf-8")
     if bcrypt.checkpw(pwd_str, cached_password_str):
         # Remove the temporary reset password from the cache.
-        reset_password_cache_key = utils.hmac_digest(
-            hmac_secret, ACCOUNT_RESET_PASSWORD_CACHE_KEY.format(username)
-        )
-        cache.delete(reset_password_cache_key)
-        cache.set(cache_key, hashed_password, ttl=cache_ttl)
+        delete_cached_reset_password(username, request.registry)
+        cache_account(hashed_password, username, request.registry)
         return True